Move AstError into separate module
