+++ crates/yuck/src/lib.rs
@@ -7,3 +7,4 @@ pub mod error;
 pub mod format_diagnostic;
 pub mod parser;
 pub mod value;
+pub mod ast_error;
